Split up tests (roughly) by prism submodule
